refactor(article,draft): integrate draft readmodel to article readmodel
